# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub Comercial iSolutis — API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_clientes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClienteResumo"
                  },
                  "type": "array",
                  "title": "Response Listar Api V1 Clientes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_clientes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/clientes/{id_}": {
      "put": {
        "tags": [
          "Clientes"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_clientes__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_clientes__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/{id_}/relacionados": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Relacionados",
        "operationId": "relacionados_api_v1_clientes__id___relacionados_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteRelacionados"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/despesas": {
      "get": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_despesas_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ano"
            }
          },
          {
            "name": "mes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DespesaLeitura"
                  },
                  "title": "Response Listar Api V1 Despesas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_despesas_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DespesaEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DespesaLeitura"
                  },
                  "title": "Response Criar Api V1 Despesas Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/despesas/opcoes": {
      "get": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Opcoes",
        "operationId": "opcoes_api_v1_despesas_opcoes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpcoesDespesa"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/despesas/resumo": {
      "get": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Resumo",
        "operationId": "resumo_api_v1_despesas_resumo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 2000,
              "title": "Ano"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoDespesas"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/despesas/{id_}": {
      "put": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_despesas__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DespesaAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DespesaLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_despesas__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/despesas/{id_}/pagar": {
      "post": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Pagar",
        "operationId": "pagar_api_v1_despesas__id___pagar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DespesaLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investimentos": {
      "get": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Listar Investimentos",
        "operationId": "listar_investimentos_api_v1_investimentos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ano"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestimentoLeitura"
                  },
                  "title": "Response Listar Investimentos Api V1 Investimentos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Criar Investimento",
        "operationId": "criar_investimento_api_v1_investimentos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestimentoEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestimentoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/investimentos/{id_}": {
      "put": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Atualizar Investimento",
        "operationId": "atualizar_investimento_api_v1_investimentos__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestimentoAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestimentoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Despesas e investimentos"
        ],
        "summary": "Excluir Investimento",
        "operationId": "excluir_investimento_api_v1_investimentos__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClienteAtualizar": {
        "properties": {
          "cpf_cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpf Cnpj"
          },
          "nome": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nome"
          },
          "segmento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmento"
          },
          "contato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Site",
                  "Indicação",
                  "LinkedIn",
                  "Instagram",
                  "WhatsApp",
                  "Evento",
                  "Prospecção ativa",
                  "Outro"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Obs"
          },
          "endereco": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Endereco"
          },
          "cep": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cep"
          },
          "municipio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipio Id"
          },
          "versao": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "versao"
        ],
        "title": "ClienteAtualizar"
      },
      "ClienteEntrada": {
        "properties": {
          "cpf_cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpf Cnpj"
          },
          "nome": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nome"
          },
          "segmento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmento"
          },
          "contato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Site",
                  "Indicação",
                  "LinkedIn",
                  "Instagram",
                  "WhatsApp",
                  "Evento",
                  "Prospecção ativa",
                  "Outro"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Obs"
          },
          "endereco": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Endereco"
          },
          "cep": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cep"
          },
          "municipio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipio Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome"
        ],
        "title": "ClienteEntrada",
        "description": "Cliente: o tipo de pessoa é deduzido do documento (11 dígitos = PF; senão PJ)."
      },
...
      "ClienteResumo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "tipo_pessoa": {
            "type": "string",
            "title": "Tipo Pessoa"
          },
          "cpf_cnpj": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpf Cnpj"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "segmento": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmento"
          },
          "contato": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contato"
          },
          "cargo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cargo"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "origem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origem"
          },
          "obs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Obs"
          },
          "endereco": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endereco"
          },
          "cep": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cep"
          },
          "municipio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipio Id"
          },
          "municipio_nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Municipio Nome"
          },
          "uf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uf"
          },
          "papeis": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Papeis"
          },
          "negocios_abertos": {
            "type": "integer",
            "title": "Negocios Abertos"
          },
          "faturado": {
            "type": "number",
            "title": "Faturado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "tipo_pessoa",
          "cpf_cnpj",
          "nome",
          "segmento",
          "contato",
          "cargo",
          "telefone",
          "email",
          "origem",
          "obs",
          "endereco",
          "cep",
          "municipio_id",
          "municipio_nome",
          "uf",
          "papeis",
          "negocios_abertos",
          "faturado"
        ],
        "title": "ClienteResumo",
        "description": "Linha da lista de clientes, com os números que a tabela mostra."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}